feat: more tempates for ts

+++ index.js
@@ -7,7 +7,7 @@ const chalk = require("chalk")
 const figlet = require("figlet")
 const ora = require('ora')
 
-const projectMap = require('./config')
+const config = require('./config')
 const version = require('./package.json').version
 
 const run = async () => {
@@ -17,17 +17,22 @@ const run = async () => {
     create(answers)
 }
 
-const print = () => console.log(
-        chalk.green(
-            figlet.textSync("YIAI-CLI", {
-                font: "Ghost",
-                horizontalLayout: "default",
-                verticalLayout: "default"
-            })
-        )
+const print = () => {
+    console.log(chalk.green(
+        figlet.textSync("YIAI-CLI", {
+            font: "Ghost",
+            horizontalLayout: "default",
+            verticalLayout: "default"
+        }))
     )
+    console.log(chalk.green('you can see more templates at: https://github.com/lq782655835/ts-templates'))
+}
 
 const askQuestions = () => {
+    const maxCharLength = Math.max(...Object.keys(config).map(key => key.length))
+    const strPadEndLength = maxCharLength + 10
+    const choices = Object.keys(config).map(key => `${key.padEnd(strPadEndLength)} ${config[key].description}`)
+
     const questions = [
         {
             name: "project",
@@ -38,19 +43,22 @@ const askQuestions = () => {
             type: "list",
             name: "tpl",
             message: "choice your project template?",
-            choices: Object.keys(projectMap)
+            choices
         }
     ]
     return inquirer.prompt(questions)
 }
 
 const create = ({project, tpl}) => {
+    let dest = `${shell.pwd()}/${project}`
+    if (shell.test('-e', dest)) {
+        console.log(chalk.red('目标文件夹已存在'))
+        process.exit(1)
+    }
+
     const spinner = ora('downloading template')
     spinner.start()
-
-    let pwd = shell.pwd()
-    let dest = `${pwd}/${project}`
-    clone(projectMap[tpl], dest, null, function() {
+    clone(config[tpl].url, dest, null, function() {
         spinner.stop()
         shell.rm('-rf', dest + `/.git`)
         console.log(chalk.white.bgGreen.bold(`Done！project created at ${dest}`))
